Extract render.py from legacy.py
